CODE/Proj2 Document Summarizer/requirements.txt
streamlit
python-docx
PyPDF2
python-dotenv
